requirements.txt
aiohttp==3.9.3
aioschedule==0.5.2
aiosignal==1.3.1
altgraph==0.17.4
APScheduler==3.10.4
attrs==23.2.0
discord.py==2.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
motor==3.3.2
multidict==6.0.5
packaging==23.2
pefile==2023.2.7
pip==24.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
pymongo==4.6.2
pytz==2024.1
pywin32-ctypes==0.2.2
setuptools==69.1.1
six==1.16.0
tzdata==2024.1
tzlocal==5.2
yarl==1.9.4
load-dotenv==0.1.0